Added buildings health growth
